+++ tt-train/sources/examples/grpo/tests/weight_transfer/configurations/local2/mgd.textproto
