--- tt-train/sources/examples/grpo_speedup/conftest.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import sys
-import os
-
-sys.path.insert(0, os.path.dirname(__file__))
